requirements.txt
attr==0.3.2
attrs==22.1.0
boto3==1.26.9
botocore==1.29.9
crowd_kit==1.2.0
pandas==1.5.1
prefect==1.4.0
requests==2.28.1
toloka_kit==0.1.23
toloka_prefect==0.0.3
